Move components out of the __init__ folder and create a components directory
