+++ splashgen/components/__init__.py
@@ -1,0 +1,2 @@
+from .SplashSite import SplashSite
+from .CTAButton import CTAButton
